Extract AbstractContextAwareExecutorService
