+++ core/src/main/java/com/linecorp/armeria/common/DefaultContextAwareExecutorService.java
@@ -15,106 +15,21 @@
  */
 package com.linecorp.armeria.common;
 
-import java.util.Collection;
-import java.util.List;
-import java.util.concurrent.Callable;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Future;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
-import java.util.stream.Collectors;
 
 import com.google.common.base.MoreObjects;
 
-class DefaultContextAwareExecutorService implements ContextAwareExecutorService {
-
+class DefaultContextAwareExecutorService extends AbstractContextAwareExecutorService {
     private final RequestContext context;
-    private final ExecutorService executor;
 
     DefaultContextAwareExecutorService(RequestContext context, ExecutorService executor) {
+        super(executor);
         this.context = context;
-        this.executor = executor;
     }
 
     @Override
     public final RequestContext context() {
         return context;
-    }
-
-    @Override
-    public final void shutdown() {
-        executor.shutdown();
-    }
-
-    @Override
-    public final List<Runnable> shutdownNow() {
-        return executor.shutdownNow();
-    }
-
-    @Override
-    public Future<?> submit(Runnable task) {
-        return executor.submit(context().makeContextAware(task));
-    }
-
-    @Override
-    public <T> Future<T> submit(Runnable task, T result) {
-        return executor.submit(context().makeContextAware(task), result);
-    }
-
-    @Override
-    public <T> Future<T> submit(Callable<T> task) {
-        return executor.submit(context().makeContextAware(task));
-    }
-
-    @Override
-    public final boolean isShutdown() {
-        return executor.isShutdown();
-    }
-
-    @Override
-    public final boolean isTerminated() {
-        return executor.isTerminated();
-    }
-
-    @Override
-    public final boolean awaitTermination(long timeout, TimeUnit unit) throws InterruptedException {
-        return executor.awaitTermination(timeout, unit);
-    }
-
-    @Override
-    public final <T> List<Future<T>> invokeAll(
-            Collection<? extends Callable<T>> tasks) throws InterruptedException {
-        return executor.invokeAll(makeContextAware(tasks));
-    }
-
-    @Override
-    public final <T> List<Future<T>> invokeAll(
-            Collection<? extends Callable<T>> tasks, long timeout, TimeUnit unit) throws InterruptedException {
-        return executor.invokeAll(makeContextAware(tasks), timeout, unit);
-    }
-
-    @Override
-    public final <T> T invokeAny(Collection<? extends Callable<T>> tasks)
-            throws InterruptedException, ExecutionException {
-        return executor.invokeAny(makeContextAware(tasks));
-    }
-
-    @Override
-    public final <T> T invokeAny(
-            Collection<? extends Callable<T>> tasks, long timeout, TimeUnit unit)
-            throws InterruptedException, ExecutionException, TimeoutException {
-        return executor.invokeAny(makeContextAware(tasks), timeout, unit);
-    }
-
-    @Override
-    public final void execute(Runnable command) {
-        executor.execute(context().makeContextAware(command));
-    }
-
-    private <T> Collection<? extends Callable<T>> makeContextAware(
-            Collection<? extends Callable<T>> tasks) {
-        return tasks.stream().map(context()::makeContextAware).collect(Collectors.toList());
     }
 
     @Override
@@ -124,6 +39,5 @@ class DefaultContextAwareExecutorService implements ContextAwareExecutorService 
                           .add("executor", executor)
                           .toString();
     }
-
 }
