+++ core/src/main/resources/com/linecorp/armeria/public_suffixes.txt
@@ -4804,6 +4804,7 @@ me.eu.org
 me.it
 me.ke
 me.so
+me.ss
 me.tc
 me.tz
 me.uk
@@ -7005,6 +7006,7 @@ sch.ng
 sch.qa
 sch.sa
 sch.so
+sch.ss
 sch.zm
 schaeffler
 schlesisches.museum
